pyproject.toml
[tool.poetry]
name = "boardgame-sommelier"
version = "0.1.0"
description = "Difyのチャットフローを利用したボードゲーム推薦チャットボットアプリケーション"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.0.2"
requests = "^2.31.0"
python-dotenv = "^1.0.1"
gunicorn = "^21.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api" 